court-processor/courtlistener_integration/requirements.txt
# CourtListener Integration Requirements
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
prometheus-client>=0.19.0
tenacity>=8.2.0  # For retry logic
ratelimit>=2.2.1  # Additional rate limiting